GROK-18428: FileEditors: LaTeX viewer
Added debugging code
